+++ src/lib.rs
@@ -31,11 +31,11 @@ pub mod service;
 pub mod storage;
 
 #[doc(inline)]
-pub use manta_util::serde;
-#[doc(inline)]
 pub use manta_accounting;
 #[doc(inline)]
 pub use manta_pay;
+#[doc(inline)]
+pub use manta_util::serde;
 
 #[doc(inline)]
 pub use tokio;